requirements.txt
aiohappyeyeballs             2.4.4
aiohttp                      3.9.5
aiosignal                    1.3.2
altair                       5.5.0
annotated-types              0.7.0
anyio                        4.8.0
attrs                        24.3.0
audioread                    3.0.1
blinker                      1.9.0
cachetools                   5.5.0
certifi                      2024.12.14
cffi                         1.17.1
charset-normalizer           3.4.1
click                        8.1.8
colorama                     0.4.6
comtypes                     1.4.9
dataclasses-json             0.6.7
datasets                     3.2.0
decorator                    5.1.1
dill                         0.3.8
distro                       1.9.0
filelock                     3.16.1
filetype                     1.2.0
frozenlist                   1.5.0
fsspec                       2024.9.0
gitdb                        4.0.12
GitPython                    3.1.44
google-ai-generativelanguage 0.6.10
google-api-core              2.24.0
google-api-python-client     2.157.0
google-auth                  2.37.0
google-auth-httplib2         0.2.0
google-generativeai          0.8.3
googleapis-common-protos     1.66.0
greenlet                     3.1.1
grpcio                       1.69.0
grpcio-status                1.69.0
gTTS                         2.5.4
h11                          0.14.0
httpcore                     1.0.7
httplib2                     0.22.0
httpx                        0.28.1
httpx-sse                    0.4.0
huggingface-hub              0.27.1
idna                         3.10
Jinja2                       3.1.5
jiter                        0.8.2
joblib                       1.4.2
jsonpatch                    1.33
jsonpointer                  3.0.0
jsonschema                   4.23.0
jsonschema-specifications    2024.10.1
langchain                    0.3.14
langchain-community          0.3.14
langchain-core               0.3.29
langchain-google-genai       2.0.8
langchain-openai             0.2.14
langchain-pinecone           0.2.0
langchain-text-splitters     0.3.5
langsmith                    0.2.10
lazy_loader                  0.4
librosa                      0.10.2.post1
llvmlite                     0.43.0
markdown-it-py               3.0.0
MarkupSafe                   3.0.2
marshmallow                  3.24.1
mdurl                        0.1.2
mpmath                       1.3.0
msgpack                      1.1.0
multidict                    6.1.0
multiprocess                 0.70.16
mypy-extensions              1.0.0
narwhals                     1.21.1
networkx                     3.4.2
numba                        0.60.0
numpy                        1.26.4
openai                       1.59.4
orjson                       3.10.13
outcome                      1.3.0.post0
packaging                    24.2
pandas                       2.2.3
pillow                       11.1.0
pinecone-client              5.0.1
pinecone-plugin-inference    1.1.0
pinecone-plugin-interface    0.0.7
pip                          24.3.1
platformdirs                 4.3.6
pooch                        1.8.2
propcache                    0.2.1
proto-plus                   1.25.0
protobuf                     5.29.2
pyarrow                      18.1.0
pyasn1                       0.6.1
pyasn1_modules               0.4.1
pycparser                    2.22
pydantic                     2.10.4
pydantic_core                2.27.2
pydantic-settings            2.7.1
pydeck                       0.9.1
Pygments                     2.19.1
pyparsing                    3.2.1
PyPDF2                       3.0.1
pypiwin32                    223
PySocks                      1.7.1
python-dateutil              2.9.0.post0
python-dotenv                1.0.1
pyttsx3                      2.98
pytz                         2024.2
pywin32                      308
PyYAML                       6.0.2
referencing                  0.35.1
regex                        2024.11.6
requests                     2.32.3
requests-toolbelt            1.0.0
rich                         13.9.4
rpds-py                      0.22.3
rsa                          4.9
safetensors                  0.5.1
scikit-learn                 1.6.0
scipy                        1.15.0
selenium                     4.27.1
sentence-transformers        3.3.1
setuptools                   65.5.0
six                          1.17.0
smmap                        5.0.2
sniffio                      1.3.1
sortedcontainers             2.4.0
soundfile                    0.13.0
soxr                         0.5.0.post1
SQLAlchemy                   2.0.36
streamlit                    1.41.1
sympy                        1.13.1
tenacity                     9.0.0
threadpoolctl                3.5.0
tiktoken                     0.8.0
tokenizers                   0.21.0
toml                         0.10.2
torch                        2.5.1
tornado                      6.4.2
tqdm                         4.67.1
transformers                 4.47.1
trio                         0.28.0
trio-websocket               0.11.1
typing_extensions            4.12.2
typing-inspect               0.9.0
tzdata                       2024.2
uritemplate                  4.1.1
urllib3                      2.3.0
watchdog                     6.0.0
websocket-client             1.8.0
wsproto                      1.2.0
xxhash                       3.5.0
yarl                         1.18.3